+++ packages/vite-plugin/src/index.ts
@@ -9,6 +9,8 @@ import {
   compile,
   IdentifierOption,
   addFileScope,
+  getPackageInfo,
+  CompileOptions,
 } from '@vanilla-extract/integration';
 import { PostCSSConfigResult, resolvePostcssConfig } from './postcss';
 
@@ -17,14 +19,22 @@ const styleUpdateEvent = (fileId: string) =>
 
 interface Options {
   identifiers?: IdentifierOption;
+  esbuildOptions?: CompileOptions['esbuildOptions'];
 }
-export function vanillaExtractPlugin({ identifiers }: Options = {}): Plugin {
+export function vanillaExtractPlugin({
+  identifiers,
+  esbuildOptions,
+}: Options = {}): Plugin {
   let config: ResolvedConfig;
   let server: ViteDevServer;
   let postCssConfig: PostCSSConfigResult | null;
   const cssMap = new Map<string, string>();
 
   let virtualExt: string;
+  let packageName: string;
+
+  const getAbsoluteVirtualFileId = (source: string) =>
+    normalizePath(path.join(config.root, source));
 
   return {
     name: 'vanilla-extract',
@@ -49,6 +59,7 @@ export function vanillaExtractPlugin({ identifiers }: Options = {}): Plugin {
     },
     async configResolved(resolvedConfig) {
       config = resolvedConfig;
+      packageName = getPackageInfo(config.root).name;
 
       if (config.command === 'serve') {
         postCssConfig = await resolvePostcssConfig(config);
@@ -56,50 +67,62 @@ export function vanillaExtractPlugin({ identifiers }: Options = {}): Plugin {
 
       virtualExt = `.vanilla.${config.command === 'serve' ? 'js' : 'css'}`;
     },
-    resolveId(id) {
-      if (!id.endsWith(virtualExt)) {
+    resolveId(source) {
+      const [validId, query] = source.split('?');
+      if (!validId.endsWith(virtualExt)) {
         return;
       }
 
-      const normalizedId = id.startsWith('/') ? id.slice(1) : id;
+      // Absolute paths seem to occur often in monorepos, where files are
+      // imported from outside the config root.
+      const absoluteId = source.startsWith(config.root)
+        ? source
+        : getAbsoluteVirtualFileId(validId);
 
-      if (cssMap.has(normalizedId)) {
-        return normalizePath(path.join(config.root, normalizedId));
+      // There should always be an entry in the `cssMap` here.
+      // The only valid scenario for a missing one is if someone had written
+      // a file in their app using the .vanilla.js/.vanilla.css extension
+      if (cssMap.has(absoluteId)) {
+        // Keep the original query string for HMR.
+        return absoluteId + (query ? `?${query}` : '');
       }
     },
     load(id) {
-      if (!id.endsWith(virtualExt)) {
+      const [validId] = id.split('?');
+
+      if (!cssMap.has(validId)) {
         return;
       }
 
-      const cssFileId = id.slice(config.root.length + 1);
-      const css = cssMap.get(cssFileId);
+      const css = cssMap.get(validId);
 
       if (typeof css !== 'string') {
         return;
       }
 
-      if (!server) {
+      if (!server || server.config.isProduction) {
         return css;
       }
 
       return outdent`
         import { injectStyles } from '@vanilla-extract/css/injectStyles';
-        
+
         const inject = (css) => injectStyles({
-          fileScope: ${JSON.stringify({ filePath: cssFileId })},
+          fileScope: ${JSON.stringify({ filePath: validId })},
           css
         });
 
         inject(${JSON.stringify(css)});
 
-        import.meta.hot.on('${styleUpdateEvent(cssFileId)}', (css) => {
+        import.meta.hot.on('${styleUpdateEvent(validId)}', (css) => {
           inject(css);
-        });   
+        });
       `;
     },
     async transform(code, id, ssrParam) {
-      if (!cssFileFilter.test(id)) {
+      const [validId] = id.split('?');
+
+      if (!cssFileFilter.test(validId)) {
         return null;
       }
 
@@ -111,37 +134,37 @@ export function vanillaExtractPlugin({ identifiers }: Options = {}): Plugin {
         ssr = ssrParam?.ssr;
       }
 
-      const index = id.indexOf('?');
-      const validId = index === -1 ? id : id.substring(0, index);
-
       if (ssr) {
         return addFileScope({
           source: code,
           filePath: normalizePath(validId),
           rootPath: config.root,
+          packageName,
         });
       }
 
       const { source, watchFiles } = await compile({
         filePath: validId,
         cwd: config.root,
+        esbuildOptions,
       });
 
       for (const file of watchFiles) {
         // In start mode, we need to prevent the file from rewatching itself.
         // If it's a `build --watch`, it needs to watch everything.
-        if (config.command === 'build' || file !== id) {
+        if (config.command === 'build' || file !== validId) {
           this.addWatchFile(file);
         }
       }
 
-      return processVanillaFile({
+      const output = await processVanillaFile({
         source,
         filePath: validId,
         identOption:
           identifiers ?? (config.mode === 'production' ? 'short' : 'debug'),
         serializeVirtualCssPath: async ({ fileScope, source }) => {
-          const id = `${fileScope.filePath}${virtualExt}`;
+          const rootRelativeId = `${fileScope.filePath}${virtualExt}`;
+          const absoluteId = getAbsoluteVirtualFileId(rootRelativeId);
 
           let cssSource = source;
 
@@ -157,26 +180,43 @@ export function vanillaExtractPlugin({ identifiers }: Options = {}): Plugin {
             cssSource = postCssResult.css;
           }
 
-          if (server && cssMap.has(id) && cssMap.get(id) !== source) {
+          if (
+            server &&
+            cssMap.has(absoluteId) &&
+            cssMap.get(absoluteId) !== source
+          ) {
             const { moduleGraph } = server;
-            const module = moduleGraph.getModuleById(id);
+            const [module] = Array.from(
+              moduleGraph.getModulesByFile(absoluteId) || [],
+            );
 
             if (module) {
               moduleGraph.invalidateModule(module);
+
+              // Vite uses this timestamp to add `?t=` query string automatically for HMR.
+              module.lastHMRTimestamp =
+                (module as any).lastInvalidationTimestamp || Date.now();
             }
 
             server.ws.send({
               type: 'custom',
-              event: styleUpdateEvent(id),
+              event: styleUpdateEvent(absoluteId),
               data: cssSource,
             });
           }
 
-          cssMap.set(id, cssSource);
+          cssMap.set(absoluteId, cssSource);
 
-          return `import "${id}";`;
+          // We use the root relative id here to ensure file contents (content-hashes)
+          // are consistent across build machines
+          return `import "${rootRelativeId}";`;
         },
       });
+
+      return {
+        code: output,
+        map: { mappings: '' },
+      };
     },
   };
 }